+++ index.php
@@ -70,7 +70,8 @@ return [
      */
     'config' => [
 
-        'logo_contrast' => ''
+        'logo_contrast' => '',
+        'logo_offcanvas' => ''
 
     ],